requirements.txt
# Core AI/ML Libraries
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
google-generativeai>=0.3.0

# Web Research
tavily-python>=0.3.0
duckduckgo-search>=4.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
markdown>=3.5.0

# Optional: For better formatting
rich>=13.7.0
